Conflicts:
	src/main/java/com/devbliss/gpullr/service/github/GithubApi.java

+++ src/main/java/com/devbliss/gpullr/repository/UserRepository.java
@@ -9,8 +9,7 @@ import org.springframework.stereotype.Component;
  * Persistence layer for {@link com.devbliss.gpullr.domain.User} objects.
  */
 @Component
-public interface UserRepository extends CrudRepository<User, Long> {
+public interface UserRepository extends CrudRepository<User, Integer> {
 
   List<User> findAll();
-
 }
